fix: 로직과 UI를 분리
