+++ src/meetDetail/hooks/useMeetStore.tsx
@@ -6,15 +6,15 @@ export const useFetchMeet = () => {
 
   const fetchMeet = async (meetId: number) => {
     try {
-      const token = localStorage.getItem("accessToken");
-      if (!token) {
-        console.log("로그인이 필요합니다.");
-        resetMeet();
-        return;
-      }
+      const token = localStorage.getItem("accessToken") || "";
+      // if (!token) {
+      //   console.log("로그인이 필요합니다.");
+      //   resetMeet();
+      //   return;
+      // }
 
-      // const res = await getMeetDetailService(meetId, token);
-      const res = await getMeetDetailService(meetId);
+      const res = await getMeetDetailService(meetId, token);
+      // const res = await getMeetDetailService(meetId);
       setMeet(res.data);
     } catch {
       console.log("모임 상세 조회에 실패했어요.");
